画面回転時にスクロール位置を保存する。EntityのIDの型をLongとStringで切り替えられるようにする。misskey疑似アカウントでプロフカラムを見れる

+++ app/src/main/java/jp/juggler/subwaytooter/table/ContentWarning.kt
@@ -74,7 +74,8 @@ object ContentWarning :TableCompanion{
 			
 			val cv = ContentValues()
 			cv.put(COL_HOST, status.hostAccessOrOriginal)
-			cv.put(COL_STATUS_ID, status.idAccessOrOriginal)
+			cv.put(COL_STATUS_ID, status.idAccessOrOriginal.toLong())
+			// TODO: Misskey用には別のテーブルを作るべき
 			cv.put(COL_SHOWN, is_shown.b2i())
 			cv.put(COL_TIME_SAVE, now)
 			App1.database.replace(table, null, cv)
